Decomplect the slack logging by putting it directly into the ingestion job. #1564
