--- ingestion/monitoring/Dockerfile
@@ -1,74 +1,0 @@
-# `python-base` sets up all our shared environment variables
-FROM python:3.10-slim as python-base
-
-ENV PYTHONUNBUFFERED=1 \
-    # prevents python creating .pyc files
-    PYTHONDONTWRITEBYTECODE=1 \
-    \
-    PIP_NO_CACHE_DIR=off \
-    PIP_DISABLE_PIP_VERSION_CHECK=on \
-    PIP_DEFAULT_TIMEOUT=100 \
-    \
-    # https://python-poetry.org/docs/configuration/#using-environment-variables
-    POETRY_VERSION=1.1.5 \
-    # make poetry install to this location
-    POETRY_HOME="/opt/poetry" \
-    # make poetry create the virtual environment in the project's root
-    # it gets named `.venv`
-    POETRY_VIRTUALENVS_IN_PROJECT=true \
-    # do not ask any interactive question
-    POETRY_NO_INTERACTION=1 \
-    \
-    # this is where our requirements + virtual environment will live
-    PYSETUP_PATH="/opt/pysetup" \
-    VENV_PATH="/opt/pysetup/.venv"
-
-# prepend poetry and venv to path
-ENV PATH="$POETRY_HOME/bin:$VENV_PATH/bin:$PATH"
-
-# `builder-base` stage is used to build deps + create our virtual environment
-FROM python-base as builder-base
-RUN apt-get update \
-    && apt-get install --no-install-recommends -y \
-        curl \
-        build-essential
-
-# install poetry - respects $POETRY_VERSION & $POETRY_HOME
-RUN curl -sSL https://install.python-poetry.org/ | python3 - --version 1.1.13
-
-# copy project requirement files here to ensure they will be cached.
-WORKDIR $PYSETUP_PATH
-COPY poetry.lock pyproject.toml ./
-
-ENV PATH="${PATH}:/root/.poetry/bin"
-
-# install runtime deps - uses $POETRY_VIRTUALENVS_IN_PROJECT internally
-RUN poetry install --no-dev
-
-# `development` image is used during development / testing
-FROM python-base as development
-
-RUN apt-get update && apt-get upgrade -y curl \
-      awscli
-
-WORKDIR $PYSETUP_PATH
-
-# copy in our built poetry + venv
-COPY --from=builder-base $POETRY_HOME $POETRY_HOME
-COPY --from=builder-base $PYSETUP_PATH $PYSETUP_PATH
-ENV PATH="${PATH}:/root/.poetry/bin"
-
-# will become mountpoint of our code
-WORKDIR /app
-
-COPY errorLogsToSlack.py poetry.lock pyproject.toml ./
-
-# quicker install as runtime deps are already installed
-RUN poetry install
-
-# notice I haven't set the environment variable values or args here.
-# the slack webhook and AWS region should be configured in the job definition,
-# and the args should be configured at submission time
-ENV AWS_REGION $AWS_REGION
-ENV SLACK_WEBHOOK $SLACK_WEBHOOK
-CMD ["poetry", "run", "python3", "./errorLogsToSlack.py"]
